+++ app/src/debug/java/io/github/droidkaigi/confsched2017/di/HttpClientModule.java
@@ -1,0 +1,34 @@
+package io.github.droidkaigi.confsched2017.di;
+
+import com.facebook.stetho.okhttp3.StethoInterceptor;
+
+import android.content.Context;
+
+import java.io.File;
+
+import javax.inject.Singleton;
+
+import dagger.Module;
+import dagger.Provides;
+import okhttp3.Cache;
+import okhttp3.Interceptor;
+import okhttp3.OkHttpClient;
+
+@Module
+public class HttpClientModule {
+
+    static final String CACHE_FILE_NAME = "okhttp.cache";
+
+    static final long MAX_CACHE_SIZE = 4 * 1024 * 1024;
+
+    @Singleton
+    @Provides
+    public OkHttpClient provideHttpClient(Context context, Interceptor interceptor) {
+        File cacheDir = new File(context.getCacheDir(), CACHE_FILE_NAME);
+        Cache cache = new Cache(cacheDir, MAX_CACHE_SIZE);
+
+        OkHttpClient.Builder c = new OkHttpClient.Builder().cache(cache).addInterceptor(interceptor)
+                .addInterceptor(new StethoInterceptor());
+        return c.build();
+    }
+}
