+++ app/src/main/java/io/github/droidkaigi/confsched2017/di/AppComponent.java
@@ -6,7 +6,7 @@ import javax.inject.Singleton;
 import dagger.Component;
 
 @Singleton
-@Component(modules = {AppModule.class, AndroidModule.class})
+@Component(modules = {AppModule.class, AndroidModule.class, HttpClientModule.class})
 public interface AppComponent {
 
     ActivityComponent plus(ActivityModule module);
